Take a command to run in the VM as positional arguments
This implements the possibility of requesting a given to be run inside
the VM with a command line like this:

    $ tuxrun [ OPTIONS ] run-this-command

If the desired command needs options, you have the separate the tuxrun
options from the command to run with --, so that those don't get
confused with options for tuxrun itself. e.g.:

    $ tuxrun [OPTIONS] -- sh -c 'command1 && command 2'

Of course, the command must be available inside the rootfs. The user has
a few options here: 1) inject the necessary files via --overlay or 2)
use a custom rootfs.

Any command specified like this is run *in addition* to any tests
specified via --tests, and after those tests have executed.

Fixes: https://gitlab.com/Linaro/tuxrun/-/issues/21