commands: enable user-defined 'commands-name'
Today when submitting a 'command' test, the suite will be named
'commands' and the test name will be 'command'.
Make it possible to pass in a user-defined 'command name' via
'--parameters commands-name=ltp-dio18', default the name to 'command'.
This makes it easier for users to define their own name depending on
what command they run.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>