gitlab-ci: build a tuxrun-dispatcher container
Add a tuxrun-dispatcher container with lava-dispatcher as the base plus podman.
This will make it cleaner when people want to run different
'--qemu-image' versions.
Today this cmdline is needed:
tuxrun --runtime podman --device qemu-arm64 --image tuxrun:fvp \
       --qemu-image docker.io/linaro/tuxrun-qemu:master

Which also mean that I have to clone tuxrun and build the fvp container.

With this patch applied, the command will be cleaner.
tuxrun --runtime podman --device qemu-arm64 \
       --qemu-image docker.io/linaro/tuxrun-qemu:master

And no need to clone and build the fvp container.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>