tests: kselftest: add kselftest overlay first
If a settings file is added via a '--overlay
https://path/to/settings.tar.xz
/opt/kselftests/default-in-kernel/arm64/'
and the KSELFTEST variable is passed via the '--parameters' command then
KSELFTEST tarball will be appended to the end of all the overlays.
This means that the settings file is unpacked and then the
kselftest tarball, that will overwrite the settings file.
This patch adds the kselftest tarball first, then if anyone
want's to add an overlay where they tweak the timeout settings files for
instance or anything else that gets overlayed into the rootfs into the
kselftest directory that will have an effect.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>