tests: perf: add perf overlay first
If a file is added via a '--overlay
https://path/to/settings.tar.xz
 That is used by perf.
and the PERF variable is passed via the '--parameters' command then PERF
tarball will be appended to the end of all the overlays. This in turn
means that if the file from the overlay is unpacked and then the perf
tarball, which means that the file will be overwritten.
If the perf tarball is added first, then if anyone
want's to add an overlay where they tweak anything in perf's tarball
files then it can be accomplished.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>